@@ -97,6 +97,7 @@ import java.util.regex.Pattern;
                 //                result.put("comname",
                 //                    this.clientCompanyRepository.findByComid(con.getComid()).getComname());
                 result.put("cstarttime", con.getCstarttime());
+                String a= con.getCstarttime().toString();
                 result.put("cendtime", con.getCendtime());
             } catch (Exception e) {
                 result.clear();
@@ -105,7 +106,8 @@ import java.util.regex.Pattern;
             }
             results.add(result);
         }
-        return results;
+//        return results;
+        return contracts;
     }
 
     //    @Override public Object sortContractList(String startOrEnd){
@@ -196,14 +198,15 @@ import java.util.regex.Pattern;
         String cname;
         String cendtime;
         String appreason;
-        int apptype;
+        int pid;
         Double camt;
         String cstarttime;
-        if (map.get("cname") != null && map.get("apptype") != null
-            &&  map.get("camt") != null &&  map.get("cstarttime") != null
-            && map.get("cendtime") != null &&  map.get("appreason") != null) {
+        if (map.get("camt") != null && map.get("cstarttime") != null
+            &&  map.get("cendtime") != null &&  map.get("cname") != null
+            && map.get("appreason") != null &&  map.get("auditstatid") != null
+            &&  map.get("suppid") != null &&  map.get("cstat") != null
+            &&  map.get("pid") != null &&  map.get("appusrid") != null ) {
             cname = (String) map.get("cname");
-            apptype = Integer.parseInt( map.get("apptype").toString());
             camt = Double.parseDouble(map.get("camt").toString());
             String amtstr =  map.get("camt").toString();
             Boolean strResult = amtstr.matches("^[0-9]+\\.?[0-9]?[0-9]?$");
@@ -213,56 +216,57 @@ import java.util.regex.Pattern;
             cstarttime =  map.get("cstarttime").toString();
             cendtime =  map.get("cendtime").toString();
             appreason =  map.get("appreason").toString();
+            pid = (int) map.get("pid");
         } else {
             return result = "表单不完整";
         }
 
-        Pattern p = Pattern.compile("^(\\d{4})/(\\d{2})/(\\d{2})$");
-        Matcher matcher1 = p.matcher(cstarttime);
-        Matcher matcher2 = p.matcher(cendtime);
-        boolean rs1 = matcher1.matches();
-        boolean rs2 = matcher2.matches();
+//        Pattern p = Pattern.compile("^(\\d{4})/(\\d{2})/(\\d{2})$");
+//        Matcher matcher1 = p.matcher(cstarttime);
+//        Matcher matcher2 = p.matcher(cendtime);
+//        boolean rs1 = matcher1.matches();
+//        boolean rs2 = matcher2.matches();
 
 
-        if (rs1 == false || rs2 == false) {
-            return "日期格式不正确";
-        }
-        int comid = Integer.parseInt((String) map.get("comid"));
-        Boolean isremind = Boolean.parseBoolean((String) map.get("isremind"));
-
+//        if (rs1 == false || rs2 == false) {
+//            return "日期格式不正确";
+//        }
+//        int comid = Integer.parseInt((String) map.get("comid"));
+//        Boolean isremind = Boolean.parseBoolean((String) map.get("isremind"));
+//
         Contract contract = new Contract();
-        Date createtime = DateUtil.getNowDate();
-        SimpleDateFormat sdf0 = new SimpleDateFormat("yyyyMMddHHmmss");
-        Random random = new Random();
-        //        contract.setCnum(sdf0.format(createtime)+ random.nextInt(10));
-        contract.setCname(cname);
-        //        contract.setCtype(ctype);
-        contract.setCamt(camt);
-        SimpleDateFormat sdf1 = new SimpleDateFormat("yyyy/MM/dd");
-        Date cstarttime1;
-        Date cendtime1;
-        try {
-            cstarttime1 = sdf1.parse(cstarttime);
-            contract.setCstarttime(cstarttime1);
-            cendtime1 = sdf1.parse(cendtime);
-            contract.setCendtime(cendtime1);
-        } catch (ParseException e1) {
-            result = "时间处理失败";
-            e1.printStackTrace();
-        }
+//        Date signdate = DateUtil.getNowDate();
+//        SimpleDateFormat sdf0 = new SimpleDateFormat("yyyyMMddHHmmss");
+//        Random random = new Random();
+//        //        contract.setCnum(sdf0.format(signdate)+ random.nextInt(10));
+//        contract.setCname(cname);
+//        //        contract.setCtype(ctype);
+//        contract.setCamt(camt);
+//        SimpleDateFormat sdf1 = new SimpleDateFormat("yyyy/MM/dd");
+//        Date cstarttime1;
+//        Date cendtime1;
+//        try {
+//            cstarttime1 = sdf1.parse(cstarttime);
+////            contract.setCstarttime(cstarttime1);
+////            cendtime1 = sdf1.parse(cendtime);
+////            contract.setCendtime(cendtime1);
+//        } catch (ParseException e1) {
+//            result = "时间处理失败";
+//            e1.printStackTrace();
+//        }
 
-        if ((contract.getCendtime().compareTo(contract.getCstarttime())) == 1) {
+//        if ((contract.getCendtime().compareTo(contract.getCstarttime())) == 1) {
             //            contract.setCremarks(cremarks);
             //            contract.setComid(comid);
             //            contract.setIsremind(isremind);
             //            contract.setCreatetime(createtime);
-            contract.setModtime(createtime);
+//            contract.setModtime(createtime);
             //            contract.setEmpid(this.comempRepository.findByUid(uid).getEmpid());
             //            contract.setBid(this.comempRepository.findByUid(uid).getBid());
 
-        } else {
-            return result = "请查看日期是否准确";
-        }
+//        } else {
+//            return result = "请查看日期是否准确";
+//        }
 
         try {
             if (null != fnames && 0 != fnames.size()) {
@@ -271,7 +275,7 @@ import java.util.regex.Pattern;
                 //            if (null != fnames && 0 != fnames.size()) {
                 for (int i = 0; i < fnames.size(); i++) {
                     int fid = this.fileService
-                        .addFile(fnames.get(i), comid, GlobalConstants.CONTRACT, appreason);
+                        .addFile(fnames.get(i), pid, GlobalConstants.CONTRACT, appreason);
                     System.out.println("fid = " + fid);
                     this.cFileService.addCFile(cid, fid);
                 }
@@ -325,7 +329,7 @@ import java.util.regex.Pattern;
             try {
                 cstarttime1 = sdf1.parse(cstarttime);
 
-                contract.setCstarttime(cstarttime1);
+//                contract.setCstarttime(cstarttime1);
             } catch (ParseException e1) {
                 result = "时间处理失败";
                 e1.printStackTrace();
@@ -335,7 +339,7 @@ import java.util.regex.Pattern;
             try {
                 cstarttime1 = sdf1.parse(cstarttime);
 
-                contract.setCstarttime(cstarttime1);
+//                contract.setCstarttime(cstarttime1);
 
             } catch (ParseException e1) {
                 result = "时间处理失败";
